adding api call and sqs integeration test

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/justicedataagentclientapi/service/integration/JdaWorkerServiceImpl.kt
@@ -1,6 +1,8 @@
 package uk.gov.justice.digital.hmpps.justicedataagentclientapi.service.integration
 
+import kotlinx.coroutines.runBlocking
 import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import software.amazon.awssdk.services.sqs.model.DeleteMessageRequest
@@ -11,16 +13,20 @@ import uk.gov.justice.digital.hmpps.justicedataagentclientapi.service.event.JdaM
 import uk.gov.justice.digital.hmpps.justicedataagentclientapi.service.integration.client.JdaWorkerClient
 import uk.gov.justice.digital.hmpps.justicedataagentclientapi.service.integration.dto.request.JdaRequest
 import uk.gov.justice.digital.hmpps.justicedataagentworker.dto.response.JdaResponse
+import uk.gov.justice.digital.hmpps.justicedataagentworker.exception.NotFoundException
 import uk.gov.justice.hmpps.sqs.HmppsQueueService
 
 @Service
 class JdaWorkerServiceImpl(
   private val jdaWorkerClient: JdaWorkerClient,
-  private val jdaMessagePublisher: JdaMessagePublisher,
-  private val hmppsQueueService: HmppsQueueService,
   private val objectMapper: ObjectMapper,
-  @param:Value("\${hmpps.sqs.queues.jdarequestqueus.queuename}") val queueName: String,
+  @param:Value("\${hmpps.sqs.queues.jdarequestqueus.queuename}") val jdaRequestQueueName: String,
+  @param:Value("\${hmpps.sqs.queues.jdaresponsequeus.queuename}") val jdaResponseQueueName: String,
 ) : JdaWorkerService {
+  @Autowired
+  private lateinit var jdaMessagePublisher: JdaMessagePublisher
+  @Autowired
+  private lateinit var hmppsQueueService: HmppsQueueService
   companion object {
     val logger = LoggerFactory.getLogger(this::class.java)
   }
@@ -32,42 +38,38 @@ class JdaWorkerServiceImpl(
   }
 
   override suspend fun dequeueResponse(): JdaResponse {
-    logger.info("Dequeue  response")
-    val sqsClient = hmppsQueueService
-      .findByQueueId("jdarequestqueus")!!.sqsClient
-    val messages = sqsClient.receiveMessage(
-      ReceiveMessageRequest.builder()
-        .maxNumberOfMessages(1)
-        .queueUrl(
-          sqsClient.getQueueUrl(
-            GetQueueUrlRequest.builder()
-              .queueName("")
-              .build(),
-          ).resultNow().queueUrl(),
+    try {
+      logger.info("Dequeue  response")
+      val sqsClient = hmppsQueueService
+        .findByQueueId("jdaresponsequeus")!!.sqsClient
+      val queueUrl = sqsClient.getQueueUrl(
+        GetQueueUrlRequest.builder()
+          .queueName(jdaResponseQueueName)
+          .build(),
+      )?.join()?.queueUrl()
+      val messages = sqsClient.receiveMessage(
+        ReceiveMessageRequest.builder()
+          .maxNumberOfMessages(1)
+          .queueUrl(queueUrl)
+          .build(),
+      )?.join()
+      if (messages?.hasMessages() == true) {
+        val jdaResponse = objectMapper.readValue(messages?.messages()[0]?.body(), JdaResponse::class.java)
+        logger.info("Deleting message from the jda response queue")
+        sqsClient.deleteMessage(
+          DeleteMessageRequest.builder()
+            .queueUrl(queueUrl)
+            .receiptHandle(messages?.messages()[0]?.receiptHandle())
+            .build(),
         )
-        .build(),
-    )
-    sqsClient.getQueueUrl(
-      GetQueueUrlRequest.builder()
-        .queueName("")
-        .build(),
-    ).resultNow().queueUrl()
-    val jdaResponse = objectMapper.readValue(messages.resultNow().messages()[0].body(), JdaResponse::class.java)
-    val message = messages.resultNow().messages()[0]
-    // Dequeue message
-    sqsClient.deleteMessage(
-      DeleteMessageRequest.builder()
-        .queueUrl(
-          sqsClient.getQueueUrl(
-            GetQueueUrlRequest.builder()
-              .queueName(queueName)
-              .build(),
-          ).resultNow().queueUrl(),
-        )
-        .receiptHandle(message.receiptHandle())
-        .build(),
-    )
-    logger.info("returning dequeued response")
-    return jdaResponse
+        logger.info("returning dequeued response")
+        return jdaResponse
+      }
+      throw NotFoundException("Queue is empty, no message in queue to dequeue")
+    } catch (e: Exception) {
+      logger.error("Error during dequeue response, queue is already empty", e)
+      throw NotFoundException("Queue is empty, no message in queue to dequeue")
+    }
+
   }
 }
